# contracts/openapi/yomora-api.yaml
openapi: 3.1.0
info:
  title: Yomora API
  version: 1.0.0
  description: API do MVP Yomora para constância de leitura, biblioteca e comunidade.
servers:
  - url: http://localhost:8080/api/v1
    description: Ambiente local
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Criar conta e iniciar sessão
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201': {description: Conta criada, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '400': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /auth/login:
    post:
      tags: [Auth]
      summary: Entrar com e-mail e senha
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses:
        '200': {description: Sessão criada, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '401': {$ref: '#/components/responses/Problem'}
  /auth/apple:
    post:
      tags: [Auth]
      summary: Entrar ou criar conta com Sign in with Apple
      security: []
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AppleCredential'}}}
      responses:
        '201': {description: Conta autenticada, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /auth/apple/events:
    post:
      tags: [Auth]
      summary: Receber notificações assinadas da Apple sobre a conta
      security: []
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [payload], properties: {payload: {type: string}}}}}
      responses:
        '204': {description: Evento processado}
        '401': {$ref: '#/components/responses/Problem'}
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotacionar refresh token
      security: []
      requestBody: {$ref: '#/components/requestBodies/RefreshToken'}
      responses:
        '200': {description: Novo par, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '401': {$ref: '#/components/responses/Problem'}
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revogar refresh token
      requestBody: {$ref: '#/components/requestBodies/RefreshToken'}
      responses:
        '204': {description: Sessão encerrada}
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Solicitar recuperação sem enumerar contas
      security: []
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
      responses:
        '202': {description: Solicitação aceita}
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Definir nova senha com token de uso único
      security: []
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [token, newPassword], properties: {token: {type: string}, newPassword: {type: string, minLength: 8, maxLength: 72}}}}}
      responses:
        '204': {description: Senha alterada}
        '401': {$ref: '#/components/responses/Problem'}
...
  /auth/email-verification/confirm:
    post:
      tags: [Auth]
      summary: Confirmar e-mail com token de uso único
      security: []
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [token], properties: {token: {type: string}}}}}
      responses:
        '204': {description: E-mail confirmado}
        '400': {$ref: '#/components/responses/Problem'}
    get:
      tags: [Auth]
      summary: Confirmar e-mail a partir do link recebido
      security: []
      parameters:
        - {name: token, in: query, required: true, schema: {type: string}}
      responses:
        '200': {description: Página de confirmação, content: {text/html: {schema: {type: string}}}}
        '400': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    Problem:
      description: Erro padronizado
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  requestBodies:
    Login:
      required: true
      content: {application/json: {schema: {type: object, required: [email, password], properties: {email: {type: string, format: email}, password: {type: string}}}}}
    RefreshToken:
      required: true
      content: {application/json: {schema: {type: object, required: [refreshToken], properties: {refreshToken: {type: string}}}}}
...
  schemas:
    Problem:
      type: object
      properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, detail: {type: string}, timestamp: {type: string, format: date-time}, errors: {type: object, additionalProperties: {type: string}}}
    RegisterRequest:
      type: object
      required: [name, username, email, password]
      properties: {name: {type: string, maxLength: 120}, username: {type: string, minLength: 3, maxLength: 40}, email: {type: string, format: email}, password: {type: string, minLength: 8, maxLength: 72}}
    TokenPair:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties: {accessToken: {type: string}, refreshToken: {type: string}, tokenType: {type: string, const: Bearer}, expiresIn: {type: integer}}
    AppleCredential:
      type: object
      required: [identityToken, authorizationCode, nonce]
      properties:
        identityToken: {type: string}
        authorizationCode: {type: string}
        nonce: {type: string, maxLength: 256}
        fullName: {type: [string, 'null'], maxLength: 120}
...